+++ CHANGELOG.md
@@ -7,6 +7,8 @@ https://github.com/coleifer/peewee/releases
 
 ## master
 
+* Fix anonymous sub-select keeping a stale `id()`-based hash after `clone()`.
+
 [View commits](https://github.com/coleifer/peewee/compare/4.2.4...master)
 
 ## 4.2.4